move tachyon.client.table to tachyon-client
