+++ core/src/main/java/tachyon/client/RemoteBlockReader.java
@@ -34,7 +34,7 @@ public interface RemoteBlockReader {
     public static RemoteBlockReader createRemoteBlockReader(TachyonConf conf) {
       try {
         return CommonUtils.createNewClassInstance(conf.getClass(Constants.USER_REMOTE_BLOCK_READER,
-            ServerConstants.USER_REMOTE_BLOCK_READER_CLASS), null, null);
+            ClientConstants.USER_REMOTE_BLOCK_READER_CLASS), null, null);
       } catch (Exception e) {
         throw Throwables.propagate(e);
       }
